OCTO-11510 Split monolithic pycaption/sami.py into pycaption/sami/ package   (constants, parser, reader, writer). Emit VTT  as class   attribute instead of leaking into inline CSS, add text-align from   Layout alignment, produce valid stylesheet rules for VTT class styles,   filter non-CSS keys, silently drop writing direction. Fix rules-dict   mutation in _recreate_style_block, inline dead hooks, add docstrings.
